Test commit to see if commit hooks work.
So, let me know if you get this alright.

When writing a commit message,
the first line becomes the subject of the message, subsequent lines are the body,
and they should be followed by a diff.

This way we can review code (at least diffs) before pushing to the main branch.
You may need to add the scripts in .git/hooks explicitly to have this work in your
branch. Specifically pre-commit, commit-msg, and post-commit. I should also add
a prefix to the subject line to make it easy for rules in thunderbird or whatever to
move these (probably annoying) messages to a separate folder.