"Finished" mutexes.
Made everything consistent with the newest version (mutex4), and swapped it in.
Hard to say if it really works yet, but I think it does.

Wrote cmpxchg, but simics traps to an illegal instruction exception, so I don't
know if it is usable yet (or at all).  It can safely be removed with
minimal (almost nonexistent) overhead in terms of mutex_lock run time, but I
don't see how I would do something like mutex_try_lock without it.

Added stubs for cond.c.