Changed lock free list to use rwlocks
Now the vast majority of time there will be no blocking. There's no guarantee
of bounded waiting, but I don't think it can do much better.

I'm not completely opposed to the idea of using thread local storage, but only
if we are guaranteed by a TA that it's ok to assume there's a hard limit on the
size of the main stack. I can't find any reference to this limit in the code,
so it doesn't seem like we should rely on it.