Last minute changes from Friday.
Plus a README file to placate github.

I don't know yet if github expects push's to it's branch or is "watching" for changes to
the original branch we are using - anyway, set up a github account and I'll add you.

We may need to "reset" and point new local branches at the github one (or somehow redirect
the original branches).