Explicitly set up context for context_switch
We can avoid mucking around with a stack crawler if we do it this way.
The unfortunate side effect is that it's tailor made for our implementation
of handler_wrapper.def and context_switch.S

There's a problem with address space copying. For some reason NULL is staying direct mapped
even though I explicitly map it and invalidate the page. Still investigating.