Added some documentation. Changed mutex to use a queue instead of bakery
I also got rid of the timer context switch flag. Its really the
same thing as disabling interrupts, but instead of just delaying a context
switch (timer interrupts are maskable), it would cause a context switch
to be dropped. I'll add some more doc to the rest of the project tonight.