We build - time to debug mutexes.
In order to get things to build, I moved all of the major types in our kernel to the same spot "kernel_types.h" which should alleviate some headaches, despite not being very scalable. In any case, the types that we have in place now are at least half of what I expect we'll need in the end.

The first thing our kernel does is fail an assertion on mutex state.
Let me know if you want me to push.