Sleep draft.
This doesn't work yet.  The problem I ran into is well summed up as - what do we do when we can't do anything?

If everyone is blocked or sleeping we would currently relaunch "init," but this is obviously not the right thing to do.

There also might be a bad race condition on launching new processes, but I need to explore it more. What I saw before I artificially inserted a new thread into my test was a mutex trying to yield, and the run queue being empty.