Initial complete version of readline
Condition variables are very simple objects designed to block
only one thread at a time. Read line currently fails because the
only thread blocks, causing the scheduler to load a new copy,
which immediately blocks, etc. Next step is to change kernel_main
and the scheduler as discussed so we have a thread to spin on
while everyone else is blocked.