Some doc - more to follow
Made v_strcpy and v_memcpy wrappers for a copy function. I don't
think I'll be making the linked sleep heap though. It is much
trickier to get or maintain the last guy in the heap, making any
space savings not worth the overhead of following tons of pointers
through randomly scattered memory.