Broke new_pages before submission!!
Nooooooo!

I should have done this earlier today, but I ran through the hurdle tests, and new_pages was broken by the test in kvm_request not unlocking its mutex if we run out of addressable space. Since we never attempt more than one big allocation in any of the other tests it didn't come up.

The only thing to do is dump our late days for the one line of code it takes to fix the problem.

I'm sorry,

Justin