Changes to loader, fork, exec, and user space exception handler.  RFC
If you don't mind I think we should start using the inline commenting feature on github to review code.  We've both spotted suspicious things in each others code in the past, and hopefully that way we can catch mistakes early. Having the diff in messages is nice for a glance at the changes but doesn't compare to the diff generated by github.

Let me know if you think that might be overkill.

- Justin